pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepwiki-cli"
version = "0.2.9"
description = "AI-powered wiki generator and code Q&A for git repositories"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12.0",
  "litellm>=1.40.0",
  "chromadb>=0.5.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
]

[project.scripts]
deepwiki = "deepwiki.cli.app:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
deepwiki = ["config/*.json"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
